Addressing comment about this class not being a constant. see https://github.com/bancorprotocol/fastlane-bot/pull/399#discussion_r1535340526
